dspy-training/requirements.txt
dspy-ai>=2.5.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
